Remove pandas, numpy dependencies
Added a helper function to import and process the CSV file containing the data. This removes the need for pandas and numpy as imports/dependencies.

These are now trimmed down to just spacy and fastAPI.

Also removed the need for the spacy model, since we can just use the basic English one; we are not doing anything else with it yet.