Added templating for web-version
Got the templates working for a human-readable/web-based access point.

Moved the `chrono` endpoint to `api` and repurposed `chrono` for the web/human-version. Keeping it as `chrono` means that later we can add `litho` or similar, while `api` can give back a variety of things at once in one dictionary.